int to numeric

+++ db/migrate/20210320154307_add_score_to_wine.rb
@@ -1,5 +1,5 @@
 class AddScoreToWine < ActiveRecord::Migration[6.1]
   def change
-    add_column :wines, :score, :integer, default: 0, limit: 100
+    add_column :wines, :score, :numeric, default: 0, limit: 100
   end
 end
